# backend-MyKP/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyKP API Documentation",
        "description": "API documentation for the MyKP backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profile/{user}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get a user's profile and KP category progress summary",
                "operationId": "177fa94c02add9deecf901cc420a273b",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "UserID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Andi Tubagus Faatih Keane"
                                                },
                                                "nim": {
                                                    "type": "string",
                                                    "example": "1234567890"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "atubagus@student.ciputra.ac.id"
                                                },
                                                "jurusan": {
                                                    "description": "Study program code; null for admins",
                                                    "type": "string",
                                                    "example": "IMT",
                                                    "nullable": true
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "student"
                                                },
                                                "profile_picture": {
                                                    "type": "string",
                                                    "example": "https://i.pravatar.cc/300?u=ken.student",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "kp_categories": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ProgressID (per-user category row)",
                                                        "type": "integer",
                                                        "example": 7
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Talkshow (Wajib BMA)"
                                                    },
                                                    "current": {
                                                        "type": "integer",
                                                        "example": 6
                                                    },
                                                    "target": {
                                                        "type": "integer",
                                                        "example": 6
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "Completed",
                                                        "enum": [
                                                            "Completed",
                                                            "On Progress"
                                                        ]
                                                    },
                                                    "percentage": {
                                                        "type": "integer",
                                                        "example": 100
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "kp_summary": {
                                            "properties": {
                                                "total_categories": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "completed": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "in_progress": {
                                                    "type": "integer",
                                                    "example": 7
                                                },
                                                "total_current": {
                                                    "type": "integer",
                                                    "example": 36
                                                },
                                                "total_target": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "overall_percentage": {
                                                    "type": "integer",
                                                    "example": 36
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum API Token",
                "bearerFormat": "string",
                "scheme": "bearer"
            }
        }
    },
...
}